backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
langchain==0.1.0
langchain-google-genai==1.0.1
langchain-community==0.0.10
chromadb==0.4.18
pypdf==4.0.1
beautifulsoup4==4.12.2
python-multipart==0.0.6
pydantic==2.5.0
